==> 5_continuous_BMI_Covariate_Included.Rmd <==
```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r include=FALSE}
rm(list = ls())
if(T){
  library(rnhanesdata)
  library(tidyverse)
  library(magrittr)
  library(data.table)
  library(randomForest)
  library(caret)
  library(keras)
  library(tensorflow)
}
```

```{r}
load(file = 'pcscore.rda')
pcscore = pcscore %>% select(-mortstat) %>%
  inner_join(Covariate_D[,c('SEQN','BMI')],by = 'SEQN')

pcscore = pcscore %>% filter(10<BMI & BMI<50)
pcscore = pcscore %>% mutate('log' = log(pcscore$BMI +1))

if(F){save(pcscore,file = 'BMI.rda')}
load(file = 'BMI.rda')

```

```{r}
y = pcscore %>% select(-SEQN)
y = y %>% mutate('log' = log(BMI+1)) %>% select(-BMI)
ysub = y[,c(1:501,which(colnames(y) == 'log'))]

set.seed(100)
trainidx = sample(nrow(y),0.7*nrow(y))

fit = lm( log ~., data = ysub, subset = trainidx)
# summary(fit)
yPred = exp(predict(fit,ysub[-trainidx,]))-1
result = cbind(yPred,yTrue = pcscore[-trainidx,'BMI']) %>% 
  na.omit() %>% as.data.frame()


# MSE of yPred and yTrue
mean((result[,1]-result[,2])^2)

# visualization
library(ggplot2)
ggplot(data=result , aes(x = 1:dim(result)[1])) + 
  geom_line(aes(y = yPred),color = 'red') + 
  geom_line(aes(y = yTrue),color = 'blue') + 
  labs(x = 'Patients', y = 'BMI',
        title  = "First 500 PCs included")
  

ggplot(data = result) + 
  aes(x = yTrue, y = yPred) +
  geom_line()
```


```{r}
# first 5 PCs + Age + Gender
ysub = pcscore[,c(1:6,which(colnames(pcscore) == 'log'))] %>% 
  left_join(Covariate_D[,c('SEQN','RIDAGEYR','Gender','Race','BMI_cat')],by = 'SEQN') %>%
  select(-SEQN)

ysub = as.data.frame(lapply(ysub, function (x) if (is.factor(x)) unclass(x) %>% as.factor() else x))
# 2 for female ,1 for male


# fit = lm(log ~ PC1 + PC2 + PC3  + PC4 + PC5 + RIDAGEYR + Gender, data = ysub, subset = trainidx)
fit = lm(log ~ .  - BMI_cat,data = ysub, subset = trainidx)
summary(fit)


yPred = exp(predict(fit,ysub[-trainidx,]))-1
result = cbind(yPred,yTrue = pcscore[-trainidx,'BMI']) %>% 
  na.omit() %>% as.data.frame()
# MSE of yPred and yTrue
mean((result[,1]-result[,2])^2)

ggplot(data=result , aes(x = 1:dim(result)[1])) + 
  geom_line(aes(y = yPred),color = 'red') + 
  geom_line(aes(y = yTrue),color = 'blue') +
  labs(x = 'Patients', y = 'BMI',
       title  = "First 5 PCs/Age/Gender included")


ggplot(data = result) + 
  aes(x = yTrue, y = yPred) +
  geom_line()

```
