==> 6_BMI_compilation.Rmd <==


```{r include=F,warning=FALSE,message=FALSE}
# rm(list = ls())

library(rnhanesdata)
library(tidyverse)
library(magrittr)
library(data.table)
library(randomForest)
library(caret)
library(keras)
library(tensorflow)
library(caTools)
library(mgcv)
library(boot)
library(car)
library(sjPlot)
library(gridExtra)
# library(PCAtools)

# https://wwwn.cdc.gov/Nchs/Nhanes/2005-2006/PAXRAW_D.htm
```




# pre processing
```{r include=F}
rm(list = ls())
keep_inx <- exclude_accel(act = PAXINTEN_D, flags = Flags_D)
accel_good_D <- PAXINTEN_D[keep_inx,] 
flags_good_D <- Flags_D[keep_inx,]


MIN_name = grep('MIN',colnames(accel_good_D),value = T)
MINdata = data.frame(SEQN = accel_good_D[,'SEQN'],accel_good_D[,MIN_name] * flags_good_D[,MIN_name]) %>% 
  group_by(SEQN) %>% summarise_all("mean") %>% as.data.frame() %>% na.omit()



SEQN_7days = data.frame(SEQN = accel_good_D[,'SEQN'],accel_good_D[,MIN_name] * flags_good_D[,MIN_name]) %>% group_by(SEQN) %>% summarise(num = n()) %>% filter(num==7) %>% select(SEQN)
MINdata = MINdata %>% filter(SEQN %in% SEQN_7days$SEQN)

if(F){
  pca_model = prcomp(MINdata %>% select(-SEQN),
                     center = T,
                     scale. = T)
  save(pca_model,file= 'pca_model.rda')
}
load(file = 'pca_model.rda')

# summary(pca_model)

if(F){
eigs <- pca_model$sdev^2
cbind(
  SD = sqrt(eigs),
  Proportion = eigs/sum(eigs),
  Cumulative = cumsum(eigs)/sum(eigs))
}




analyticData = data.frame(SEQN=MINdata$SEQN,
                          pca_model$x) %>% 
  inner_join(Covariate_D %>% select(SEQN,Race,Gender,RIDAGEYR,BMI),by = 'SEQN') %>% 
  na.omit() 

analyticData = analyticData %>% filter(5 <BMI & BMI< 50)



save(MIN_name,MINdata,analyticData,file = 'analyticData.rda')

```




# Random Forest

```{r}
rm(list = ls())
load(file = 'analyticData.rda')

PCnames = paste('PC',1:5,sep = '')
y = analyticData[,c('RIDAGEYR','Race','Gender','BMI',PCnames)] %>% na.omit()





set.seed(111)
id = sample(dim(y)[1],0.8*dim(y)[1])
if(F){
  fit = randomForest(BMI ~ ., data = y,ntree = 2000,
                     subset = id,
                     importance = T)
  save(fit,file = 'rffit.rda')
}
load(file = 'rffit.rda')
library(randomForest)
varImpPlot(fit,
           main = 'Variable Importance Measured by Random Forest')

ytest = y[-id,'BMI']
ypred = fit %>% predict(y[-id,])

cor(ytest,ypred)
mean((ytest-ypred)^2)
plot(ypred,ytest)
```



# corplot
```{r}
y = analyticData[,c('RIDAGEYR','Race','Gender','BMI',PCnames)] %>% na.omit()
# importance(fit)

y_cor <- as.data.frame(
  lapply(y, function (x) if (is.factor(x)) unclass(x) %>% as.numeric  else x))

library(ggcorrplot)
ggcorrplot(cor(y_cor), 
           lab = T,
           tl.cex = 15,
           title = "Pearson’s Correlation Analysis between \n BMI, Covariates and Physical Activity") + 
  labs(caption = 'Note. RIDAGEYR - Age in years at screening.') + 
  theme(plot.title = element_text(size = 20),
        plot.caption = element_text(size = 15))

```






# PCA results visualization
```{r}
rm(list = ls())
library(gridExtra)
library(grid)
load(file = 'pca_model.rda')
load('analyticData.rda')


# scree plot
library(factoextra)
fviz_eig(pca_model)



# eigenvectors
temp = pca_model$rotation %>% data.frame()

gs = lapply(1:3, function(i) {
  ggplot(temp) + 
    aes(x = 1:nrow(temp)/60,y = unlist(temp[,i]) %>% abs) + 
    geom_line() +
    labs(x = "", y = paste0("PC",i)) +
    theme(
      plot.margin = unit(c(0, 0.1, 0, 0), "cm"),
      axis.title.y = element_text(size = rel(1.2))
          )
})

do.call(grid.arrange,c(gs,nrow = 3))




# scale = F / different percentiles
if(F){
  pca_model = prcomp(MINdata %>% select(-SEQN),
                     scale. = F)
  save(pca_model,file= 'pca_model_unscale.rda')
}




load(file = 'pca_model_unscale.rda')
pcscore = data.frame(SEQN = MINdata$SEQN, pca_model$x %>% abs)

gs = lapply(1:3, function(i) {
  id_high = pcscore$SEQN[which(pcscore[,paste0('PC',i)] == max(pcscore[,paste0('PC',i)]))]
  id_low = pcscore$SEQN[which(pcscore[,paste0('PC',i)] == min(pcscore[,paste0('PC',i)]))]
  id_25 = pcscore$SEQN[which(pcscore[,paste0('PC',i)] == quantile(pcscore[,paste0('PC',i)],c(0.25,0.5,0.75),type = 3)[1])]
  id_50 = pcscore$SEQN[which(pcscore[,paste0('PC',i)] == quantile(pcscore[,paste0('PC',i)],c(0.25,0.5,0.75),type = 3)[2])]
  id_75 = pcscore$SEQN[which(pcscore[,paste0('PC',i)] == quantile(pcscore[,paste0('PC',i)],c(0.25,0.5,0.75),type = 3)[3])]  
  tmp_high = MINdata %>% filter(SEQN %in% id_high) %>% select(-SEQN) %>% t %>% as.data.frame()
  tmp_low = MINdata %>% filter(SEQN %in% id_low) %>% select(-SEQN) %>% t %>% as.data.frame()
  # tmp_25 = MINdata %>% filter(SEQN %in% id_25) %>% select(-SEQN) %>% t %>% as.data.frame()
  tmp_50 = MINdata %>% filter(SEQN %in% id_50) %>% select(-SEQN) %>% t %>% as.data.frame()
  # tmp_75 = MINdata %>% filter(SEQN %in% id_75) %>% select(-SEQN) %>% t %>% as.data.frame()
return(
  ggplot() +
    geom_line(data = tmp_high, aes(x = 1:1440/60, y =  unlist(tmp_high), color = paste0('Subject with the Highest PC Scores'))) +
    geom_line(data = tmp_low, aes(x = 1:1440/60, y = unlist(tmp_low), color = paste0('Subject with the Lowest PC Scores'))) +
    # geom_line(data = tmp_25, aes(x = 1:1440/60, y =  unlist(tmp_25), color = paste0('Subject with the 25th percentile of PC Scores')))  +
    geom_line(data = tmp_50, aes(x = 1:1440/60, y =  unlist(tmp_50), color = paste0('Subject with the 50th percentile of PC Scores')))  +
    # geom_line(data = tmp_75, aes(x = 1:1440/60, y =  unlist(tmp_75), color = paste0('Subject with the 75th percentile of PC Scores')))  +
    labs(x='',
         y= paste0('PC',i),
         fill = "") +
    # labs(x = '', y ='') +
    theme(
      legend.title = element_blank(),
      legend.text = element_text(size = rel(0.7)),
      plot.margin = unit(c(0, 0.1, 0, 0), "cm"),
      axis.title.y = element_text(size = rel(1.2))
          )
    # scale_colour_manual(name="",values = c('black','red','green'))
)
})
grid_arrange_shared_legend <- function(...) {
    plots <- list(...)
    g <- ggplotGrob(plots[[1]] + theme(legend.position="bottom"))$grobs
    legend <- g[[which(sapply(g, function(x) x$name) == "guide-box")]]
    lheight <- sum(legend$height)
    grid.arrange(
        do.call(arrangeGrob, lapply(plots, function(x)
            x + theme(legend.position="none"))),
        legend,
        ncol = 1,
        # left = 'Averaged Daily Activity Count',
        # bottom = 'Time of Day',
        heights = grid::unit.c(unit(1, "npc") - lheight, lheight)
        )
}

do.call(grid_arrange_shared_legend,gs)




# log transform

load(file = 'pca_model_unscale.rda')
pcscore = data.frame(SEQN = MINdata$SEQN, pca_model$x %>% abs)

gs = lapply(1:3, function(i) {
  id_high = pcscore$SEQN[which(pcscore[,paste0('PC',i)] == max(pcscore[,paste0('PC',i)]))]
  id_low = pcscore$SEQN[which(pcscore[,paste0('PC',i)] == min(pcscore[,paste0('PC',i)]))]
  id_25 = pcscore$SEQN[which(pcscore[,paste0('PC',i)] == quantile(pcscore[,paste0('PC',i)],c(0.25,0.5,0.75),type = 3)[1])]
  id_50 = pcscore$SEQN[which(pcscore[,paste0('PC',i)] == quantile(pcscore[,paste0('PC',i)],c(0.25,0.5,0.75),type = 3)[2])]
  id_75 = pcscore$SEQN[which(pcscore[,paste0('PC',i)] == quantile(pcscore[,paste0('PC',i)],c(0.25,0.5,0.75),type = 3)[3])]  
  tmp_high = MINdata %>% filter(SEQN %in% id_high) %>% select(-SEQN) %>% t %>% as.data.frame()
  tmp_low = MINdata %>% filter(SEQN %in% id_low) %>% select(-SEQN) %>% t %>% as.data.frame()
  tmp_25 = MINdata %>% filter(SEQN %in% id_25) %>% select(-SEQN) %>% t %>% as.data.frame()
  tmp_50 = MINdata %>% filter(SEQN %in% id_50) %>% select(-SEQN) %>% t %>% as.data.frame()
  tmp_75 = MINdata %>% filter(SEQN %in% id_75) %>% select(-SEQN) %>% t %>% as.data.frame()
return(
  ggplot() +
    geom_line(data = tmp_high, aes(x = 1:1440/60, y =  unlist(tmp_high) %>% log(), color = paste0('Subject with the Highest PC Scores'))) +
    geom_line(data = tmp_low, aes(x = 1:1440/60, y = unlist(tmp_low) %>% log(), color = paste0('Subject with the Lowest PC Scores'))) +
    # geom_line(data = tmp_25, aes(x = 1:1440/60, y =  unlist(tmp_25) %>% log(), color = paste0('Subject with the 25th percentile of PC Scores')))  +
    geom_line(data = tmp_50, aes(x = 1:1440/60, y =  unlist(tmp_50) %>% log(), color = paste0('Subject with the 50th percentile of PC Scores')))  +
    # geom_line(data = tmp_75, aes(x = 1:1440/60, y =  unlist(tmp_75) %>% log(), color = paste0('Subject with the 75th percentile of PC Scores')))  +
    labs(x='',
         y= paste0('PC',i),
         fill = "") +
    # labs(x = '', y ='') +
    theme(
      legend.title = element_blank(),
      legend.text = element_text(size = rel(0.7)),
      plot.margin = unit(c(0, 0.1, 0, 0), "cm"),
      axis.title.y = element_text(size = rel(1.2))
          )
    # scale_colour_manual(name="",values = c('black','red','green'))
)
})

do.call(grid_arrange_shared_legend,gs)
```



# Cross-validation
```{r}
rm(list = ls())
load(file = 'analyticData.rda')


res_RMSE = c()
res_Rsquared = c()
if(F){
for(i in 1:100){
  PCnames = paste('PC',1:i,sep = '')
  train.control <- trainControl(method = "cv", number = 10)
  dat = analyticData[,c('RIDAGEYR','Race','Gender','BMI',PCnames)]
  model <- caret::train(BMI ~., data = dat, method = "lm",
               trControl = train.control)
  res_RMSE = c(res_RMSE, model$results$RMSE)
  res_Rsquared = c(res_Rsquared, model$results$Rsquared)
 
}
  save(res_RMSE,res_Rsquared, file = 'CV_100PCs.rda')
}


rm(list = ls())
load(file = 'analyticData.rda')

res_RMSE = c()
res_Rsquared = c()
if(F){
for(i in 1:50){
  PCnames = paste('PC',1:i,sep = '')
  train.control <- trainControl(method = "cv", number = 10)
  dat = analyticData[,c('RIDAGEYR','Race','Gender','BMI',PCnames)]
  model <- caret::train(BMI ~., data = dat, method = "lm",
               trControl = train.control)
  res_RMSE = c(res_RMSE, model$results$RMSE)
  res_Rsquared = c(res_Rsquared, model$results$Rsquared)
  
}
  save(res_RMSE,res_Rsquared, file = 'CV_50PCs.rda')
}

load(file = 'CV_50PCs.rda')
g1 = ggplot(data.frame(res_RMSE)) + 
  aes(x = 1:length(res_RMSE), y = res_RMSE) +
  geom_line(aes(color = 'Original RMSE Loss'))+
  # geom_smooth(se = TRUE, method = 'lm')+
  geom_smooth(se = TRUE, method =  'auto', aes(color = 'LOESS Curve Fitting')) +
  labs(x = 'Number of Principal Components Used in Cross-Validation',y = 'Root Mean Square Error ') + 
  scale_colour_manual(name="Legend", values = c("blue",'red'))

g2 = ggplot(data.frame(res_Rsquared)) + 
  aes(x = 1:length(res_Rsquared), y = res_Rsquared) +
  geom_line(aes(color = 'Original R-squared'))+
  # geom_smooth(se = TRUE, method = 'lm')+
  geom_smooth(se = TRUE, method =  'auto', aes(color = 'LOESS Curve Fitting')) + 
  labs(x = 'Number of Principal Components Used in Cross-Validation',y = 'R-squared') +
  scale_colour_manual(name="Legend", values = c("blue",'red'))

which(res_RMSE==min(res_RMSE))
which(res_Rsquared==max(res_Rsquared))
gridExtra::grid.arrange(g1,g2)

```
# ANOVA
```{r}
models = list()
for (i in 1:14){
  PCnames = paste('PC',1:i,sep = '')
  y = analyticData[,c('RIDAGEYR','Race','Gender','BMI',PCnames)] %>% na.omit()
  set.seed(111)
  trainidx = sample(nrow(y),0.7*nrow(y))
  model = lm(BMI ~., data = y, subset = trainidx)
  models = c(models,list(model))
}


do.call(anova,models)

# print(xtable(anova(model)))
```



# Linear regression


```{r}
rm(list =ls())
load(file = 'analyticData.rda')


PCnames = paste('PC',1:8,sep = '')
Y = analyticData[,c('RIDAGEYR','Race','Gender','BMI',PCnames)] %>% na.omit()
y = Y
yTrue = y$BMI
set.seed(111)
trainidx = sample(nrow(y),0.7*nrow(y))
fit2 = lm(BMI ~ .,data = y, subset = trainidx)







PCnames = paste('PC',1:14,sep = '')
Y = analyticData[,c('RIDAGEYR','Race','Gender','BMI',PCnames)] %>% na.omit()
y = Y
yTrue = y$BMI
# y = as.data.frame(lapply(y, function (x) if (is.factor(x)) unclass(x) %>% as.factor() else x))
# set.seed(111)
# trainidx = sample(nrow(y),0.7*nrow(y))


# y$BMI = log(y$BMI +1)



fit = lm(BMI ~ .,data = y, subset = trainidx)


# anova(fit,fit2)
summary(fit)
# summary(fit2)

# library(car)
# avPlots(glm(BMI ~ .,data = y, subset = trainidx),ask = FALSE)
if(F){
vif(fit)
}
# y_cor <- as.data.frame(
#   lapply(y, function (x) if (is.factor(x)) unclass(x) %>% as.numeric  else x))
# res_cor = cor(y_cor)
# library(corrplot)
# corrplot(res_cor,
#          metho = 'color',
#          type = 'upper',
#          order = 'hclust',
#          addCoef.col = "black",
#          diag = F)

# yPred = exp(predict(fit,y[-trainidx,]))-1
yPred = predict(fit,y[-trainidx,])
yTrue = yTrue[-trainidx]
result = cbind(yPred,
               yTrue) %>% as.data.frame() %>% na.omit()

# MSE of yPred and yTrue
# mean((yPred-yTrue)^2)



ggplot(data=result , aes(x = 1:dim(result)[1])) + 
  geom_line(aes(y = yPred),color = 'red') + 
  geom_line(aes(y = yTrue),color = 'blue') +
  labs(x = 'Patients', y = 'BMI',
       title  = "PCs + Age + Race + Gender")


ggplot(data = result) + 
  aes(x = yTrue, y = yPred) +
  geom_point()+
  # geom_smooth(se = TRUE, method = 'lm')+
  geom_smooth(se = TRUE, method =  'auto',color = 'red') +
  labs(y = 'Predicted BMI', x = 'Actual BMI') +
  theme(axis.text=element_text(size=20),
        axis.title=element_text(size=20))

# plot(fit2,which = 1,main = "interaction term added")
plot(fit,which = 1)




```


# Linear regression results

```{r}


# MSE
mean((result[,1]-result[,2])^2)

# cor
cor(result[,1], result[,2])

plot(yPred, yTrue)

# summary(fit2)
```



```{r}
# library(pixiedust)
# dust(fit) %>% 
#   sprinkle(cols = c("estimate", "std.error", "statistic"), round = 3) %>%
#       sprinkle(cols = "p.value", fn = quote(pvalString(value))) %>% 
#       sprinkle_colnames("Term", "Coefficient", "SE", "T-statistic", 
#                         "P-value")

```

# Plot Table
```{r message=F}
# library(sjmisc)
# library(sjlabelled)

library(sjPlot)
tab_model(fit,
          ci.hyphen = ", ",
          string.ci = "CI (95%)",
          string.p = "P-Value",
          dv.labels = "Regression Model using the first 14 PCs")
tab_model(fit2,
          ci.hyphen = ", ",
          string.ci = "CI (95%)",
          string.p = "P-Value",
          dv.labels = "Regression Model using the first 8 PCs")

tab_model(fit,fit2,
          ci.hyphen = ", ",
          show.ci = FALSE,
          dv.labels = c("Regression Model using the first 14 PCs",
                        "Regression Model using the first 8 PCs"))
```





# DEEP LEARNING data processing
```{r}
load('analyticData.rda')
library(rnhanesdata)
keep_inx <- exclude_accel(act = PAXINTEN_D, flags = Flags_D)
accel_good_D <- PAXINTEN_D[keep_inx,] 
flags_good_D <- Flags_D[keep_inx,]


MIN_name = grep('MIN',colnames(accel_good_D),value = T)
MINdata = data.frame(SEQN = accel_good_D[,'SEQN'],accel_good_D[,MIN_name] * flags_good_D[,MIN_name]) %>%
  group_by(SEQN) %>% summarise_each(funs(mean)) %>% as.data.frame() %>% na.omit()


SEQN_7days = data.frame(SEQN = accel_good_D[,'SEQN'],accel_good_D[,MIN_name] * flags_good_D[,MIN_name]) %>% group_by(SEQN) %>% summarise(num = n()) %>% filter(num==7) %>% select(SEQN)

MINdata = MINdata %>% filter(SEQN %in% SEQN_7days$SEQN)


MINdata= MINdata %>% 
  inner_join(Covariate_D %>% select(SEQN,Race,Gender,RIDAGEYR,BMI),by = 'SEQN') %>% 
  na.omit() 

# 1440*7 format

MINdata_1440 = data.frame(SEQN = accel_good_D[,'SEQN'],accel_good_D[,MIN_name] * flags_good_D[,MIN_name]) %>% as.data.frame() %>% na.omit()
MINdata_1440 = MINdata_1440 %>% filter(SEQN %in% SEQN_7days$SEQN)
MINdata_1440= MINdata_1440 %>% 
  inner_join(Covariate_D %>% select(SEQN,Race,Gender,RIDAGEYR,BMI),by = 'SEQN') %>% 
  na.omit() 

# 1440*7 = 10080 format

MINdata_10080 = data.frame('SEQN' = accel_good_D[,'SEQN'], 'WEEKDAY' = accel_good_D[,'WEEKDAY'],accel_good_D[,6:dim(accel_good_D)[2]] * flags_good_D[,6:dim(accel_good_D)[2]]) %>% 
  arrange(SEQN,WEEKDAY) %>%
  reshape(idvar = "SEQN", timevar = "WEEKDAY",direction = "wide") %>% na.omit()
MINdata_10080= MINdata_10080 %>% 
  inner_join(Covariate_D %>% select(SEQN,Race,Gender,RIDAGEYR,BMI),by = 'SEQN') %>% 
  na.omit() 

MINdata_10080 = MINdata_10080 %>% filter(SEQN %in% SEQN_7days$SEQN)

save(MINdata,MINdata_1440,MINdata_10080,file='MINdata.rda')
```







# CNN 
```{r}
rm(list = ls())
library(keras)
library(caret)
load('MINdata.rda')


MINdata_1440 = as.data.frame(lapply(MINdata_1440, function (x) if (is.factor(x)) unclass(x) %>% as.numeric else x))

# MINdata_10080 = MINdata_10080 %>% select(-Race,-Gender,-RIDAGEYR)

set.seed(000)
trainIdx = sample(c(TRUE, FALSE), dim(MINdata_1440)[1], replace = TRUE, prob = c(.7, .3))

# y = log(MINdata$BMI + 1)
y = MINdata_1440$BMI
x = MINdata_1440 %>% select(-BMI) %>% select(-SEQN)

ytrain = y[trainIdx]
xtrain = x[trainIdx, ] %>% scale()

mns = attr(xtrain, "scaled:center")
sds = attr(xtrain, "scaled:scale")

xtest = x[!trainIdx, ] %>% scale(center = mns, scale = sds)
ytest = y[!trainIdx]



# we are using univariate time series data so number of feature is 1
# for multivariate data e.g. activity signal on axis 1, axis 2, axis 3, then number of feature is 3 
xtrain = array(xtrain, dim = c(dim(xtrain)[1], dim(xtrain)[2], 1))
xtest = array(xtest, dim = c(dim(xtest)[1], dim(xtest)[2], 1))

model_CNN = keras_model_sequential() %>%
  layer_conv_1d(filters = 2^8, kernel_size = 2,
               input_shape = c(dim(xtrain)[2:3]), activation = "relu") %>%
  layer_max_pooling_1d(pool_size = 2) %>%
  layer_conv_1d(filters = 2^4, kernel_size = 2, activation = "relu") %>%
  layer_max_pooling_1d(pool_size = 2) %>%
  layer_flatten() %>%
  layer_dense(units = 2^4, activation = "relu") %>%
  # layer_dropout(0.5) %>%
  layer_dense(units = 1, activation = "linear")


model_CNN %>% compile(
 loss = "mse",
 optimizer = "adam",
 metrics = list("mean_absolute_error")
)

history = model_CNN %>% fit(
 xtrain,
 ytrain,
 epochs = 20,
 validation_split = 0.2,
 # batch_size = 100,
 verbose = 1
)

yPred = model_CNN %>% predict(xtest)
```

```{r}
model_CNN %>% save_model_hdf5("model_CNN.h5")
model_CNN <- load_model_hdf5("model_CNN.h5")
```

```{r}
plot(yPred, ytest)

mean((yPred-ytest)^2)

cor(yPred,ytest)

summary(model_CNN)
```


# CNN multi-inputs
```{r}
# CNN: separate covariates and time-series and then concatenate

rm(list = ls())
library(tidyverse)
library(reticulate)
library(keras)
library(caret)
load('MINdata.rda')


MINdata_1440 = as.data.frame(lapply(MINdata_1440, function (x) if (is.factor(x)) unclass(x) %>% as.numeric else x))

set.seed(000)
trainIdx = sample(c(TRUE, FALSE), dim(MINdata_1440)[1], replace = TRUE, prob = c(.7, .3))

# y = log(MINdata$BMI + 1)
y = MINdata_1440$BMI
x = MINdata_1440 %>% select(-BMI) %>% select(-SEQN)

ytrain = y[trainIdx]
xtrain = x[trainIdx, ] %>% scale()

mns = attr(xtrain, "scaled:center")
sds = attr(xtrain, "scaled:scale")

xtest = x[!trainIdx, ] %>% scale(center = mns, scale = sds)
ytest = y[!trainIdx]


# we are using univariate time series data so number of feature is 1
# for multivariate data e.g. activity signal on axis 1, axis 2, axis 3, then number of feature is 3 
xtrain = array(xtrain, dim = c(dim(xtrain)[1], dim(xtrain)[2], 1))
xtest = array(xtest, dim = c(dim(xtest)[1], dim(xtest)[2], 1))

# 1440 minites, feature = 1
time_input = layer_input(shape = c(1440,1),
                         name = 'input_time')

time_output = time_input %>%
  layer_conv_1d(filters = 2^8, kernel_size = 2, activation = "relu") %>%
  layer_max_pooling_1d(pool_size = 2) %>%
  layer_conv_1d(filters = 2^4, kernel_size = 2, activation = "relu") %>%
  layer_max_pooling_1d(pool_size = 2) %>%
  layer_flatten()  %>%
  layer_dense(units = 2^4, activation = "relu") 

  
covariates_input = layer_input(shape = c(3,1),
                               name = 'input_covariates')

covariates_output = covariates_input %>%
  layer_reshape(target_shape = 3)


concatenate_layer = layer_concatenate(c(time_output,covariates_output)) %>%
  layer_dense(units = 1, activation = 'linear')

model_CNN = keras_model(
  inputs = c(time_input,covariates_input),
  outputs = concatenate_layer
)

model_CNN %>% compile(
 loss = "mse",
 optimizer = "adam",
 metrics = list("mean_absolute_error")
)

# kerasR::plot_model(model_CNN,
#                    # to_file = "model.png",
#                    show_shapes = TRUE,
#   show_layer_names = TRUE)



history = model_CNN %>% fit(
  x = list(input_time = array(xtrain[,1:1440,1],dim=c(9688,1440,1)),
           input_covariates = array(xtrain[,1441:1443,1],dim=c(9688,3,1))),
  y = ytrain,
  epochs = 15,
  validation_split = 0.2,
  # batch_size = 32,
  verbose = 1
)

yPred = model_CNN %>% predict(
  list(array(xtest[1:4171,1:1440,1],dim=c(4171,1440,1)),
       array(xtest[1:4171,1441:1443,1],dim=c(4171,3,1)))
)



```

```{r}
model_CNN %>% save_model_hdf5("model_CNN.h5")
model_CNN <- load_model_hdf5("model_CNN.h5")
```

```{r}
plot(yPred, ytest)

mean((yPred-ytest)^2)

cor(yPred,ytest)

summary(model_CNN)
```




# CNN using PC scores and covariates
```{r}
rm(list = ls())
library(tidyverse)
library(reticulate)
library(keras)
library(caret)
load(file = 'analyticData.rda')


analyticData = as.data.frame(lapply(analyticData, function (x) if (is.factor(x)) unclass(x) %>% as.numeric else x))

set.seed(000)
trainIdx = sample(c(TRUE, FALSE), dim(analyticData)[1], replace = TRUE, prob = c(.7, .3))

# y = log(MINdata$BMI + 1)
y = analyticData$BMI
x = analyticData %>% select(-BMI) %>% select(-SEQN)

ytrain = y[trainIdx]
xtrain = x[trainIdx, ] %>% scale()

mns = attr(xtrain, "scaled:center")
sds = attr(xtrain, "scaled:scale")

xtest = x[!trainIdx, ] %>% scale(center = mns, scale = sds)
ytest = y[!trainIdx]



xtrain = array(xtrain, dim = c(dim(xtrain)[1], dim(xtrain)[2], 1))
xtest = array(xtest, dim = c(dim(xtest)[1], dim(xtest)[2], 1))



pc_input = layer_input(shape = c(1440,1),
                         name = 'input_pc')

pc_output = pc_input %>%
  layer_conv_1d(filters = 2^8, kernel_size = 2, activation = "relu") %>%
  layer_max_pooling_1d(pool_size = 2) %>%
  layer_conv_1d(filters = 2^4, kernel_size = 2, activation = "relu") %>%
  layer_max_pooling_1d(pool_size = 2) %>%
  layer_flatten()  %>%
  layer_dense(units = 2^4, activation = "relu") 

  
covariates_input = layer_input(shape = c(3,1),
                               name = 'input_covariates')

covariates_output = covariates_input %>%
  layer_reshape(target_shape = 3)


concatenate_layer = layer_concatenate(c(pc_output,covariates_output)) %>%
  layer_dense(units = 1, activation = 'linear')

model_CNN = keras_model(
  inputs = c(pc_input,covariates_input),
  outputs = concatenate_layer
)

model_CNN %>% compile(
 loss = "mse",
 optimizer = "adam",
 metrics = list("mean_absolute_error")
)

# kerasR::plot_model(model_CNN,
#                    # to_file = "model.png",
#                    show_shapes = TRUE,
#   show_layer_names = TRUE)



history = model_CNN %>% fit(
  x = list(input_pc = array(xtrain[,1:1440,1],dim=c(dim(xtrain)[1],1440,1)),
           input_covariates = array(xtrain[,1441:1443,1],dim=c(dim(xtrain)[1],3,1))),
  y = ytrain,
  epochs = 15,
  validation_split = 0.2,
  # batch_size = 32,
  verbose = 1
)

yPred = model_CNN %>% predict(
  list(array(xtest[,1:1440,1],dim=c(dim(xtest)[1],1440,1)),
       array(xtest[,1441:1443,1],dim=c(dim(xtest)[1],3,1)))
)


```



```{r}
plot(yPred, ytest)

mean((yPred-ytest)^2)

cor(yPred,ytest)

summary(model_CNN)
```

# LSTM muti-inputs


```{r}
rm(list = ls())
library(tidyverse)
library(reticulate)
library(keras)
library(caret)
load('MINdata.rda')


MINdata_1440 = as.data.frame(lapply(MINdata_1440, function (x) if (is.factor(x)) unclass(x) %>% as.numeric else x))

set.seed(000)
trainIdx = sample(c(TRUE, FALSE), dim(MINdata_1440)[1], replace = TRUE, prob = c(.7, .3))

# y = log(MINdata$BMI + 1)
y = MINdata_1440$BMI
x = MINdata_1440 %>% select(-BMI) %>% select(-SEQN)

ytrain = y[trainIdx]
xtrain = x[trainIdx, ] %>% scale()

mns = attr(xtrain, "scaled:center")
sds = attr(xtrain, "scaled:scale")

xtest = x[!trainIdx, ] %>% scale(center = mns, scale = sds)
ytest = y[!trainIdx]


xtrain = array(xtrain, dim = c(dim(xtrain)[1], dim(xtrain)[2], 1))
xtest = array(xtest, dim = c(dim(xtest)[1], dim(xtest)[2], 1))


time_input = layer_input(shape = c(1440,1),
                         name = 'input_time')

time_output = time_input %>%
  layer_lstm(units = 1, activation = 'relu', input_shape = c(1440,1)) %>% 
  layer_batch_normalization()

covariates_input = layer_input(shape = c(3,1),
                               name = 'input_covariates')

covariates_output = covariates_input %>%
  layer_batch_normalization() %>%
  layer_reshape(target_shape = 3) 


concatenate_layer = layer_concatenate(c(time_output,covariates_output)) %>%
  layer_dense(units = 1, activation = 'linear')

model_LSTM = keras_model(
  inputs = c(time_input,covariates_input),
  outputs = concatenate_layer
)

model_LSTM %>% compile(
 loss = "mse",
 optimizer = optimizer_adam(clipnorm=0.001),
 metrics = list("mean_absolute_error")
)

# kerasR::plot_model(model,
#                    # to_file = "model.png",
#                    show_shapes = TRUE,
#   show_layer_names = TRUE)



history = model_LSTM %>% fit(
  x = list(input_time = array(xtrain[,1:1440,1],dim=c(dim(xtrain)[1],1440,1)),
           input_covariates = array(xtrain[,1441:1443,1],dim=c(dim(xtrain)[1],3,1))),
  y = ytrain,
  epochs = 15,
  validation_split = 0.2,
  batch_size = 32,
  verbose = 1
)

yPred = model_LSTM %>% predict(
  list(array(xtest[1:4171,1:1440,1],dim=c(4171,1440,1)),
       array(xtest[1:4171,1441:1443,1],dim=c(4171,3,1)))
)

model_LSTM %>% save_model_hdf5("model_LSTM.h5")
model_LSTM <- load_model_hdf5("model_LSTM.h5")
```


```{r}
plot(yPred, ytest)

mean((yPred-ytest)^2)

cor(yPred,ytest)

summary(model_LSTM)
```

# garbage code



# CNN + LSTM

```{r}
rm(list = ls())
load('MINdata.rda')


MINdata_1440 = as.data.frame(lapply(MINdata_1440, function (x) if (is.factor(x)) unclass(x) %>% as.numeric else x))
set.seed(000)
trainIdx = sample(c(TRUE, FALSE), dim(MINdata_1440)[1], replace = TRUE, prob = c(.7, .3))


y = MINdata_1440$BMI 
x = MINdata_1440 %>% select(-BMI) %>% select(-SEQN)

ytrain = y[trainIdx]
xtrain = x[trainIdx, ] %>% scale()

mns = attr(xtrain, "scaled:center")
sds = attr(xtrain, "scaled:scale")

xtest = x[!trainIdx, ] %>% scale(center = mns, scale = sds)
ytest = y[!trainIdx]


xtrain = array(xtrain, dim = c(dim(xtrain)[1], dim(xtrain)[2], 1))
xtest = array(xtest, dim = c(dim(xtest)[1], dim(xtest)[2], 1))


model_custom = keras_model_sequential() %>%
  layer_conv_1d(filters = 2^8, kernel_size = 2,
               input_shape = c(dim(xtrain)[2:3]), activation = "relu") %>%
  layer_max_pooling_1d(pool_size = 2) %>%
  layer_lstm(units = 2^3) %>%
  layer_dense(1, activation = 'linear')

model_custom %>% compile(
 loss = "mse",
 optimizer = "adam",
 metrics = list("mean_absolute_error")
)

history = model_custom %>% fit(
 xtrain,
 ytrain,
 epochs = 20,
 validation_split = 0.2,
 verbose = 1
)

yPred = model_custom %>% predict(xtest)
```

```{r}
model_custom %>% save_model_hdf5("model_custom.h5")
model_custom <- load_model_hdf5("model_custom.h5")
```

```{r}
plot(yPred, ytest)

mean((yPred-ytest)^2)

cor(yPred,ytest)

summary(model_custom)
```

# LSTM

```{r}
rm(list = ls())
load('MINdata.rda')
MINdata_1440 = as.data.frame(lapply(MINdata_1440, function (x) if (is.factor(x)) unclass(x) %>% as.numeric else x))
set.seed(000)
trainIdx = sample(c(TRUE, FALSE), dim(MINdata_1440)[1], replace = TRUE, prob = c(.7, .3))


y = MINdata_1440$BMI 
x = MINdata_1440 %>% select(-BMI) %>% select(-SEQN)

ytrain = y[trainIdx]
xtrain = x[trainIdx, ] %>% scale()

mns = attr(xtrain, "scaled:center")
sds = attr(xtrain, "scaled:scale")

xtest = x[!trainIdx, ] %>% scale(center = mns, scale = sds)
ytest = y[!trainIdx]


tstep = 1
# we are using univariate time series data so number of feature is 1
# for multivariate data e.g. activity signal on axis 1, axis 2, axis 3, then number of feature is 3
# [samples / batch size, tstep, number of features] 
xtrain = array(xtrain, dim = c(dim(xtrain)[1], tstep, 1))
xtest = array(xtest, dim = c(dim(xtest)[1], tstep, 1))


model_LSTM = keras_model_sequential() %>%
  layer_lstm(input_shape = c(dim(xtrain)[2:3]),
             units = 2^8, activation = 'relu') %>% 
  # layer_lstm(units = 2^4, activation = 'relu') %>%
  layer_dense(2^2) %>%
  # layer_dropout(0.25) %>%
  layer_dense(units = 1, activation = "linear")


model_LSTM %>% compile(
 loss = "mse",
 optimizer = "adam",
 metrics = list("mean_absolute_error")
)

history = model_LSTM %>% fit(
 xtrain,
 ytrain,
 epochs = 10,
 validation_split = 0.2,
 verbose = 1
)

yPred = model_LSTM %>% predict(xtest)
```

```{r}
model_LSTM %>% save_model_hdf5("model_LSTM.h5")
model_LSTM <- load_model_hdf5("model_LSTM.h5")
```

```{r}
plot(yPred, ytest)

mean((yPred-ytest)^2)

cor(yPred,ytest)

summary(model_LSTM)
```


## CNN with separate covariates 
```{r}
library(keras)
library(caret)
load('MINdata.rda')


covariates = MINdata %>% select(Race,Gender,RIDAGEYR)
covariates = as.data.frame(lapply(covariates, function (x) if (is.factor(x)) unclass(x) %>% as.numeric else x))

MINdata = as.data.frame(lapply(MINdata, function (x) if (is.factor(x)) unclass(x) %>% as.numeric else x)) %>% select(-Race,-Gender,-RIDAGEYR)

# MINdata_10080 = MINdata_10080 %>% select(-Race,-Gender,-RIDAGEYR)

set.seed(000)
trainIdx = sample(c(TRUE, FALSE), dim(MINdata)[1], replace = TRUE, prob = c(.7, .3))

# y = log(MINdata$BMI + 1)
y = MINdata$BMI
x = MINdata %>% select(-BMI) %>% select(-SEQN)

ytrain = y[trainIdx]
xtrain = x[trainIdx, ] %>% scale()

mns = attr(xtrain, "scaled:center")
sds = attr(xtrain, "scaled:scale")

xtest = x[!trainIdx, ] %>% scale(center = mns, scale = sds)
ytest = y[!trainIdx]

xtrain = cbind(xtrain,covariates[trainIdx, ]) %>% as.matrix()
# xtrain = lapply(xtrain,as.numeric) %>% as.data.frame()
xtest = cbind(xtest,covariates[!trainIdx,]) %>% as.matrix()
# xtest = lapply(xtest,as.numeric) %>% as.data.frame()

# we are using univariate time series data so number of feature is 1
# for multivariate data e.g. activity signal on axis 1, axis 2, axis 3, then number of feature is 3 
xtrain = reticulate::array_reshape(xtrain, dim = c(nrow(xtrain), dim(xtrain)[2], 1))
xtest = reticulate::array_reshape(xtest, dim = c(nrow(xtest), dim(xtest)[2], 1))

model = keras_model_sequential() %>%
  layer_conv_1d(filters = 2^8, kernel_size = 2,
               input_shape = c(dim(xtrain)[2:3]), activation = "relu") %>%
  layer_max_pooling_1d(pool_size = 2) %>%
  layer_conv_1d(filters = 2^4, kernel_size = 2) %>% 
  layer_flatten() %>%
  layer_dense(units = 2^4, activation = "relu") %>%
  layer_dense(units = 1, activation = "linear")

# summary(model)

model %>% compile(
 loss = "mse",
 optimizer = "adam",
 metrics = list("mean_absolute_error")
)

history = model %>% fit(
 xtrain,
 ytrain,
 epochs = 20,
 validation_split = 0.2,
 verbose = 1,
)

yPred = model %>% predict(xtest)
```



```{r}
# yPred = exp(yPred)-1
# ytest = exp(ytest)-1

summary(model)

plot(yPred, ytest)

mean((yPred-ytest)^2)

cor(yPred,ytest)
```



# try different window size
```{r}
if(F){
err = rep(0,10)
for (d in 1:10){
MINdata_smooth = MINdata[,MIN_name] %>%
  apply(MARGIN = 2, function(i) runmean(i,k=d))
MINdata_smooth = data.frame(SEQN = MINdata$SEQN,MINdata_smooth)
pca_smooth = prcomp(MINdata_smooth %>% select(-SEQN),
                 center = T,
                 scale. = T)

y = data.frame(SEQN= MINdata_smooth$SEQN,pca_smooth$x[,1:5]) %>% 
  left_join(Covariate_D[,c('RIDAGEYR','Race','Gender','BMI','SEQN')],by = 'SEQN')
y = y[,c('RIDAGEYR','Race','Gender','BMI','PC1','PC2','PC3','PC4','PC5')] %>% na.omit()

yTrue = y$BMI

y$BMI = log(y$BMI+1)

fit = lm(BMI ~ .,data = y, subset = trainidx)
summary(fit)

yPred = exp(predict(fit,y[-trainidx,]))-1
yTrue = yTrue[-trainidx]
result = cbind(yPred,
               yTrue) %>% as.data.frame() %>% na.omit()


err[d] = mean((result[,1]-result[,2])^2)
}
err
}
# > err
#  [1] 47.93726 49.89128 50.71071 51.17472 51.32988 51.30029 51.16800 51.16298 51.10647
# [10] 51.17049
```
